requirements.txt
# 爬虫核心
requests==2.31.0  #HTTP请求核心 数据获取基础
beautifulsoup4==4.12.2 #HTML解析器 提取数据
lxml==4.9.3 #高性能解析引擎 提升HTML解析速度
selenium==4.15.0

# 数据处理
pandas==2.1.3 #数据处理 数据清洗，存储
numpy==1.24.3 #pandas依赖
jieba==0.42.1 #中文分词（推荐系统）
snownlp==0.12.3 #情感分析（推荐系统）

# 数据库
pymysql==1.1.0 #MySQL链接 数据存储
sqlalchemy==2.0.23 #数据库orm 防止sql注入

# 推荐算法
scikit-learn==1.3.2 #机器学习 余弦相似度，特征提取
surprise==0.1 #推荐系统 过滤算法，推荐算法

# Web与可视化
django==4.2.7 #Web框架 注册系统，显示界面
pyecharts==2.0.3 #数据可视化

# 工具库
python-dotenv==1.0.0 #配置管理 存储数据库敏感信息
tqdm==4.66.1 #进度显示